New approach to combo manager.
My old approach has a problem where a user doesn't have
to enter the correct combo to pull it off. The old way
only checks that the proper buttons were hit in sequence.
If the user hits a wrong button somewhere in the middle
it would not matter.

That means most of the old code has to get ripped out.

The new approach is to spy on ig.input.bind/bindTouch
and keep track of streams of input. If the last n inputs
match any of the registered combos then the combo goes
off. Simpler, ultimately.